feat: add extensions for iva and implemente anulation changes
Took 4 hours 33 minutes

+++ src/main/java/com/pawscodes/sierras/kafka/bitrix/data/entity/DocumentLinPed.java
@@ -40,7 +40,7 @@ public class DocumentLinPed {
     int porc_dcto_2 = 0;
     @Builder.Default
     int porc_dcto_3 = 0;
-    int despacho_virtual;
+    double despacho_virtual;
     double cantidad_otra_und;
     int cantidad_dos;
 }
